requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-debug-toolbar==4.4.2
django-ninja==1.1.0
django-seed==0.3.1
djangorestframework==3.15.2
Faker==25.9.1
gunicorn==22.0.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
toposort==1.10
typing_extensions==4.12.2
